+++ extra_foam/pipeline/__init__.py
@@ -1,9 +1,2 @@
-from .worker import PulseWorker, TrainWorker
-from .pipe import MpInQueue, MpOutQueue
-
-__all__ = [
-    "MpInQueue",
-    "MpOutQueue",
-    "PulseWorker",
-    "TrainWorker",
-]
+from .f_worker import PulseWorker, TrainWorker
+from .f_pipe import MpInQueue, MpOutQueue
